+++ core-ng/src/main/java/core/framework/api/kafka/MessageHandler.java
@@ -1,0 +1,9 @@
+package core.framework.api.kafka;
+
+/**
+ * @author neo
+ */
+@FunctionalInterface
+public interface MessageHandler<T> {
+    void handle(String key, T message) throws Exception;
+}