+++ core-ng/src/main/java/core/framework/impl/json/JSONReader.java
@@ -27,8 +27,12 @@ public final class JSONReader<T> {
     }
 
     public T fromJSON(byte[] json) {
+        return fromJSON(json, 0, json.length);
+    }
+
+    public T fromJSON(byte[] json, int offset, int length) {
         try {
-            return reader.readValue(json);
+            return reader.readValue(json, offset, length);
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }