+++ core-ng/src/main/java/core/framework/impl/log/stat/CollectStatTask.java
@@ -3,7 +3,7 @@ package core.framework.impl.log.stat;
 import core.framework.api.util.ASCII;
 import core.framework.api.util.Lists;
 import core.framework.api.util.Maps;
-import core.framework.impl.log.LogForwarder;
+import core.framework.impl.log.KafkaLogForwarder;
 
 import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
@@ -33,9 +33,9 @@ public class CollectStatTask implements Runnable {
     private final ThreadMXBean thread = ManagementFactory.getThreadMXBean();
     private final MemoryMXBean memory = ManagementFactory.getMemoryMXBean();
     private final List<GCStat> gcStats;
-    private final LogForwarder logForwarder;
+    private final KafkaLogForwarder logForwarder;
 
-    public CollectStatTask(LogForwarder logForwarder) {
+    public CollectStatTask(KafkaLogForwarder logForwarder) {
         this.logForwarder = logForwarder;
 
         gcStats = Lists.newArrayList();