+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ## Change log
 ### 4.7.8 (9/15/2016 - ongoing)
 * lib: update undertow to 1.4.2
+* queue: start supporting kafka
 
 ### 4.7.7 (9/14/2016 - 9/15/2016)
 * http: throw methodNotAllowedException for unknown http method