+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 
 ## Change log
 ### 4.7.8 (9/15/2016 - ongoing)
-* lib: update undertow to 1.4.2
+* lib: update undertow to 1.4.3
+* es: update esTookTime to nano seconds
+* mockito: update mockito to 2.1
 
 ### 4.7.7 (9/14/2016 - 9/15/2016)
 * http: throw methodNotAllowedException for unknown http method